requirements.txt
anthropic
sounddevice
numpy
pyaudio
webrtcvad
pyannote.audio
torch
whisperx
flask
soundfile
